Typo fix
The serialiser and deserialiser treat it as a uint16. Can't change that because it'll break the PC and mobile apps.

+++ datatypes.h
@@ -405,7 +405,7 @@ typedef struct {
 	bool multi_esc;
 	bool tc;
 	float tc_max_diff;
-	uint32_t update_rate_hz;
+	uint16_t update_rate_hz;
 } adc_config;
 
 // Nunchuk control types
